refactor(rimap-server): split wire_validator.rs into envelope/supervisor/inbound/outbound
Splits the 1564-LOC wire_validator.rs into a wire_validator/
directory module:
  - mod.rs (809): ValidatedStdio, ValidatorSupervisor (struct),
    ValidationOutcome, ErrorEnvelope, BUF_SIZE, stdio_with_validation
    factory, and the test module (37 tests covering envelope
    validation, bridge I/O, and supervisor lifecycle).
  - envelope.rs (404): pure JSON-RPC envelope validation —
    is_forwardable_id / is_valid_params / is_well_formed_error /
    extract_id / parse_error / invalid_request / the three
    dup-key visitor structs / has_duplicate_keys_in_rmcp_strict_positions
    / validate / synthesize_error_line.
  - inbound.rs (106): validate_inbound bridge task.
  - outbound.rs (45): passthrough_outbound bridge task.
  - supervisor.rs (121): impl ValidatorSupervisor (watch_for_error,
    drain, shutdown_after_failure, take_or_await, flatten).

pub(crate) re-exports preserve fuzz_oracle's
`super::wire_validator::{ValidationOutcome, synthesize_error_line,
validate}` import unchanged. Tests stay in mod.rs because they
exercise items from every submodule.

No behavior change. All 1416 workspace tests still pass.

Refs #300 (item 2)
