--- crates/rimap-server/src/mcp/wire_validator.rs
@@ -1,1564 +1,0 @@
-//! JSON-RPC 2.0 envelope validator (#277).
-//!
-//! Wraps rmcp's stdio transport: reads stdin, validates each line as a
-//! JSON-RPC 2.0 envelope, forwards valid envelopes to rmcp via an
-//! in-memory duplex stream, and synthesizes `-32600 Invalid Request` /
-//! `-32700 Parse error` envelopes for malformed input. Without this
-//! layer rmcp 1.5's `JsonRpcMessageCodec` silently drops envelopes
-//! that fail its compatibility shim — see
-//! `docs/superpowers/specs/2026-05-15-issue-277-envelope-validator-design.md`.
-
-use std::sync::Arc;
-
-use serde_json::Value;
-use tokio::io::{AsyncBufReadExt, AsyncRead, AsyncWriteExt, BufReader, DuplexStream, Stdout};
-use tokio::sync::Mutex;
-use tokio::task::JoinHandle;
-
-/// Bridge-task duplex buffer. Large enough that one well-formed
-/// envelope fits in a single write; both directions are independent
-/// tasks so inbound stalls cannot cause outbound deadlock.
-// cargo-mutants: known-equivalent — `64 * 1024` vs `64 + 1024` is a
-// buffer-size constant; no test asserts on the value and both sizes
-// are large enough to hold one envelope per write.
-const BUF_SIZE: usize = 64 * 1024;
-
-/// What the validator decides for a given line.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum ValidationOutcome {
-    /// Forward the line byte-for-byte to rmcp.
-    Forward,
-    /// Empty or whitespace-only line — drop silently.
-    Skip,
-    /// Reject with the synthesized error envelope.
-    Reject(ErrorEnvelope),
-}
-
-/// A synthesized JSON-RPC error response.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct ErrorEnvelope {
-    pub code: i32,
-    pub message: &'static str,
-    pub id: Value, // `Value::Null` if absent or invalid; otherwise echoed string/number.
-}
-
-/// Public bundle handed back to `main.rs::run`. `transport` plugs into
-/// `rmcp::serve_server(server, validated.transport)`; `stdout` is the
-/// shared writer that all synchronized output paths (validator
-/// rejections, passthrough frames, AND `emit_pre_init_error_envelope`)
-/// lock; `supervisor` exposes the bridge-task lifecycle.
-pub struct ValidatedStdio {
-    pub transport: (DuplexStream, DuplexStream),
-    pub stdout: Arc<Mutex<Stdout>>,
-    pub supervisor: ValidatorSupervisor,
-}
-
-/// Supervises the two bridge tasks (`validate_inbound` and
-/// `passthrough_outbound`). See the spec for the three-method lifecycle
-/// contract: race-phase fail-fast (`watch_for_error`), success-path
-/// drain (`drain`), and failure-path abort+drain (`shutdown_after_failure`).
-///
-/// **`JoinHandle` lifecycle.** `watch_for_error` may poll either or
-/// both `JoinHandle`s to completion. Once a `JoinHandle` has been
-/// polled to completion, re-polling panics (tokio invariant: see
-/// `tokio::runtime::task::core` line 422 — "`JoinHandle` polled after
-/// completion"). The `_consumed` flags track this so the success-path
-/// `drain` and failure-path `shutdown_after_failure` can skip the
-/// re-await on already-consumed handles. The skipped result is treated
-/// as `Ok(())` per the invariant that `drain` is only called after
-/// `watch_for_error` returned `Ok` (both bridges exited cleanly) and
-/// `shutdown_after_failure` always aborts before awaiting.
-pub struct ValidatorSupervisor {
-    pub(crate) inbound: JoinHandle<std::io::Result<()>>,
-    pub(crate) outbound: JoinHandle<std::io::Result<()>>,
-    pub(crate) inbound_consumed: bool,
-    pub(crate) outbound_consumed: bool,
-}
-
-// ============================================================
-// Pure validator (no I/O — unit-testable in isolation).
-// ============================================================
-
-/// `id` accepted by rmcp's `RxJsonRpcMessage`. rmcp 1.5's
-/// `RequestId = NumberOrString` rejects null; strings are unrestricted;
-/// numbers must be i64-representable (rejects fractional values and
-/// numbers outside i64 range — `serde_json` parses very large ints as
-/// f64 which `as_i64` also rejects).
-pub(crate) fn is_forwardable_id(v: &Value) -> bool {
-    if v.is_string() {
-        return true;
-    }
-    v.as_i64().is_some()
-}
-
-/// `params` accepted on JSON-RPC requests and notifications. JSON-RPC §4
-/// nominally allows a Structured value (Array OR Object) when present,
-/// but rmcp's `CustomRequest` / `CustomNotification` deserializers
-/// route `params` through `serde(flatten)` over a `WithMeta { _meta,
-/// _rest }` wrapper, which can only deserialize from a map shape —
-/// so an `Array` body produces "data did not match any variant of
-/// untagged enum `JsonRpcMessage`" and rmcp silently drops the line
-/// (server appears hung). The cargo-fuzz oracle (#266) caught
-/// `{"jsonrpc":"2.0","method":"x","id":1,"params":[1,2,3]}` on the
-/// seed corpus.
-///
-/// `Null` is accepted because rmcp tolerates it as "no parameters"
-/// and rejecting it would over-strict legitimate clients. Number,
-/// String, and Boolean shapes are likewise silently dropped by rmcp.
-pub(crate) fn is_valid_params(v: &Value) -> bool {
-    v.is_object() || v.is_null()
-}
-
-/// `error` body matches JSON-RPC §5.1: an object with i32-representable
-/// `code` and string `message`. `data` is optional. rmcp 1.5's
-/// `ErrorCode = i32`, so fractional values and numbers outside i32
-/// range fail rmcp's deserialization and are rejected here.
-pub(crate) fn is_well_formed_error(v: &Value) -> bool {
-    let Some(obj) = v.as_object() else {
-        return false;
-    };
-    let code_ok = obj
-        .get("code")
-        .and_then(Value::as_i64)
-        .is_some_and(|n| i32::try_from(n).is_ok());
-    let message_ok = obj.get("message").is_some_and(Value::is_string);
-    code_ok && message_ok
-}
-
-/// Read the top-level `id` field for echo on a rejection envelope.
-/// Returns `Value::Null` if `id` is missing, present-but-null, or of a
-/// disallowed type; otherwise echoes the original value verbatim.
-/// JSON-RPC §5 says the id on a synthesized error response MUST be
-/// null when the original could not be detected.
-///
-/// **`is_forwardable_id` symmetry.** Only id shapes that would have
-/// been forwardable (string OR i64-representable number) are echoed.
-/// Fractional numbers, oversized numbers, arrays, objects, and
-/// booleans fall through to `Null`. This matches MCP's `RequestId =
-/// integer | string` schema, which the synthesized error envelope
-/// must satisfy. Cargo-fuzz oracle (#266) caught `{".sonrpc":"2.0",
-/// "id":2.5}`: typo'd `jsonrpc` reaches `extract_id` before the
-/// id-validity branch, the lax echo emitted `id: 2.5`, and the MCP
-/// schema rejected the resulting envelope ("2.5 is not of types
-/// integer, string").
-pub(crate) fn extract_id(obj: &serde_json::Map<String, Value>) -> Value {
-    match obj.get("id") {
-        Some(v) if is_forwardable_id(v) => v.clone(),
-        _ => Value::Null,
-    }
-}
-
-pub(crate) fn parse_error() -> ErrorEnvelope {
-    ErrorEnvelope {
-        code: -32700,
-        message: "Parse error",
-        id: Value::Null,
-    }
-}
-
-pub(crate) fn invalid_request(id: Value) -> ErrorEnvelope {
-    ErrorEnvelope {
-        code: -32600,
-        message: "Invalid Request",
-        id,
-    }
-}
-
-/// Detects duplicates in one map level and drains all keys/values
-/// without recursing. For non-map shapes returns `false` —
-/// duplicates are a map-level concept. Module-private helper for
-/// [`has_duplicate_keys_in_rmcp_strict_positions`].
-struct OneLevelDupCheck;
-
-// cargo-mutants: known-equivalent group — the non-`visit_map` methods
-// below have several mutants the test suite does not kill, by design:
-//   * `expecting` is invoked only by serde to format error diagnostics;
-//     its return value never reaches our control flow.
-//   * `visit_string` and `visit_none` are unreachable from
-//     `serde_json::Deserializer::from_str` (which is what
-//     `has_duplicate_keys_in_rmcp_strict_positions` constructs): the
-//     streaming deserializer prefers `visit_str` for JSON strings and
-//     `visit_unit` for JSON null, never the owned-String or `Option`
-//     paths.
-//   * `visit_seq` mutated to a constant return skips the drain loop;
-//     the surrounding `DupCheckOneLevel`/`map.next_value()?` chain
-//     then leaves the parser mid-array, the outer
-//     `de.deserialize_any(...).unwrap_or(false)` swallows the
-//     resulting trailing-data error, and the dup-check signals "no
-//     duplicates" — identical to the unmutated outcome (drained,
-//     returns `Ok(false)`).
-// Inline `visit_<primitive>` mutants (i64/u64/f64/bool/unit) ARE
-// killed by `error_body_as_primitive_echoes_id` further down.
-impl<'de> serde::de::Visitor<'de> for OneLevelDupCheck {
-    type Value = bool;
-
-    fn expecting(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str("any JSON value")
-    }
-
-    fn visit_map<A: serde::de::MapAccess<'de>>(self, mut map: A) -> Result<bool, A::Error> {
-        let mut seen: std::collections::HashSet<String> = std::collections::HashSet::new();
-        let mut dup = false;
-        while let Some(key) = map.next_key::<String>()? {
-            let _: serde::de::IgnoredAny = map.next_value()?;
-            if !seen.insert(key) {
-                dup = true;
-            }
-        }
-        Ok(dup)
-    }
-
-    fn visit_seq<A: serde::de::SeqAccess<'de>>(self, mut seq: A) -> Result<bool, A::Error> {
-        while seq.next_element::<serde::de::IgnoredAny>()?.is_some() {}
-        Ok(false)
-    }
-
-    fn visit_bool<E>(self, _: bool) -> Result<bool, E> {
-        Ok(false)
-    }
-    fn visit_i64<E>(self, _: i64) -> Result<bool, E> {
-        Ok(false)
-    }
-    fn visit_u64<E>(self, _: u64) -> Result<bool, E> {
-        Ok(false)
-    }
-    fn visit_f64<E>(self, _: f64) -> Result<bool, E> {
-        Ok(false)
-    }
-    fn visit_str<E>(self, _: &str) -> Result<bool, E> {
-        Ok(false)
-    }
-    fn visit_string<E>(self, _: String) -> Result<bool, E> {
-        Ok(false)
-    }
-    fn visit_unit<E>(self) -> Result<bool, E> {
-        Ok(false)
-    }
-    fn visit_none<E>(self) -> Result<bool, E> {
-        Ok(false)
-    }
-}
-
-/// Newtype with a `Deserialize` impl that defers to
-/// [`OneLevelDupCheck`]. Lets the outer visitor invoke
-/// `map.next_value::<DupCheckOneLevel>()?` to recurse exactly once
-/// into the `error` subtree.
-struct DupCheckOneLevel(bool);
-
-impl<'de> serde::de::Deserialize<'de> for DupCheckOneLevel {
-    fn deserialize<D>(d: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        d.deserialize_any(OneLevelDupCheck).map(DupCheckOneLevel)
-    }
-}
-
-/// Detects duplicates at the top level AND inside the `error`
-/// subtree (one level deep), but does not recurse further. Module-
-/// private helper for
-/// [`has_duplicate_keys_in_rmcp_strict_positions`].
-struct TopAndErrorDupCheck;
-
-// cargo-mutants: known-equivalent group — every method below except
-// `visit_map` produces an observably-equivalent outcome under any
-// stub-return mutation. `has_duplicate_keys_in_rmcp_strict_positions`
-// is called from `validate(line)` BEFORE the line is parsed into a
-// `Value`. For a non-object top-level (string/number/bool/null/array)
-// the parsed `Value` later fails `parsed.as_object()` and is rejected
-// via `invalid_request(Value::Null)` — the same id used by the
-// dup-check rejection path. So whether `visit_<primitive>` returns
-// `Ok(true)` or `Ok(false)`, the final ValidationOutcome is the same
-// `Reject(invalid_request(Value::Null))`. `expecting` only formats
-// serde diagnostics. `visit_seq` without drain triggers a trailing-
-// data error that the outer `unwrap_or(false)` swallows back to the
-// "no duplicates" verdict — same outcome path as drained-then-
-// Ok(false). `visit_map` (the one we actually care about) IS killed
-// by `duplicate_top_level_keys_reject` and `duplicate_keys_inside_error_body_reject`.
-impl<'de> serde::de::Visitor<'de> for TopAndErrorDupCheck {
-    type Value = bool;
-
-    fn expecting(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str("any JSON value")
-    }
-
-    fn visit_map<A: serde::de::MapAccess<'de>>(self, mut map: A) -> Result<bool, A::Error> {
-        // Drain ALL keys before returning; early-return would leave
-        // the streaming deserializer's input position mid-map and
-        // `deserialize_any` would propagate a trailing-data error,
-        // surfacing as `Ok(false)` here via the outer
-        // `unwrap_or(false)`. Accumulate the dup flag instead.
-        let mut seen: std::collections::HashSet<String> = std::collections::HashSet::new();
-        let mut dup = false;
-        while let Some(key) = map.next_key::<String>()? {
-            if key == "error" {
-                // Recurse one level into `error`'s value, since
-                // rmcp's `ErrorData` is a strict struct deserialize.
-                let nested: DupCheckOneLevel = map.next_value()?;
-                if nested.0 {
-                    dup = true;
-                }
-            } else {
-                let _: serde::de::IgnoredAny = map.next_value()?;
-            }
-            if !seen.insert(key) {
-                dup = true;
-            }
-        }
-        Ok(dup)
-    }
-
-    fn visit_seq<A: serde::de::SeqAccess<'de>>(self, mut seq: A) -> Result<bool, A::Error> {
-        while seq.next_element::<serde::de::IgnoredAny>()?.is_some() {}
-        Ok(false)
-    }
-
-    fn visit_bool<E>(self, _: bool) -> Result<bool, E> {
-        Ok(false)
-    }
-    fn visit_i64<E>(self, _: i64) -> Result<bool, E> {
-        Ok(false)
-    }
-    fn visit_u64<E>(self, _: u64) -> Result<bool, E> {
-        Ok(false)
-    }
-    fn visit_f64<E>(self, _: f64) -> Result<bool, E> {
-        Ok(false)
-    }
-    fn visit_str<E>(self, _: &str) -> Result<bool, E> {
-        Ok(false)
-    }
-    fn visit_string<E>(self, _: String) -> Result<bool, E> {
-        Ok(false)
-    }
-    fn visit_unit<E>(self) -> Result<bool, E> {
-        Ok(false)
-    }
-    fn visit_none<E>(self) -> Result<bool, E> {
-        Ok(false)
-    }
-}
-
-/// Returns `true` iff `line` contains duplicate JSON keys in any
-/// position rmcp deserializes via a strict `#[derive(Deserialize)]`
-/// struct.
-///
-/// `serde_json::from_str::<Value>` silently collapses duplicates
-/// (last wins), but rmcp's `serde(untagged)` + `serde(flatten)`
-/// machinery on `JsonRpcMessage` rejects duplicate keys at strict-
-/// struct positions, so a line that looks valid after dedup gets
-/// silently dropped by rmcp. Detected with a streaming visitor on
-/// the raw input before any parse-to-`Value` step so the dedup
-/// never happens.
-///
-/// **Positions checked:**
-/// - The top-level envelope object (`jsonrpc`, `id`, `method`,
-///   `params`, `result`, `error`).
-/// - The `error` body if present (`code`, `message`, `data`) —
-///   rmcp's `ErrorData` is a strict `#[derive(Deserialize)]` struct.
-///
-/// **Positions NOT checked** (rmcp uses lenient `Value` /
-/// `JsonObject` here, so duplicates collapse last-wins and don't
-/// cause rejection): `params`, `result`, `error.data`, and any
-/// further-nested subtrees.
-///
-/// Non-object roots and unparsable input return `false` — those
-/// shapes are caught by the existing parse-error / non-object
-/// branches in `validate`.
-///
-/// Cargo-fuzz oracle (#266) findings caught by this helper:
-/// - Top level: `{"jsonrpc":"2.0","id":99,"res":"2.0","id":99,"result":{"x":1}}`.
-/// - `error` body: `{"jsonrpc":"2.0","id":1,"error":{"code":175,"message":75,"message":"x"}}`.
-fn has_duplicate_keys_in_rmcp_strict_positions(line: &str) -> bool {
-    use serde::Deserializer as _;
-    let mut de = serde_json::Deserializer::from_str(line);
-    de.deserialize_any(TopAndErrorDupCheck).unwrap_or(false)
-}
-
-/// Validate one line of input and decide what to do with it.
-pub(crate) fn validate(line: &str) -> ValidationOutcome {
-    if line.trim().is_empty() {
-        return ValidationOutcome::Skip;
-    }
-    // Detect duplicate keys in any rmcp strict-struct position
-    // (top-level envelope OR `error` body) on the raw bytes BEFORE
-    // parsing into `Value`, since `Value` collapses duplicates
-    // silently and rmcp rejects them. See
-    // `has_duplicate_keys_in_rmcp_strict_positions` docstring.
-    // Echo `Null` for the id — with duplicates, no single id value is
-    // safely echoable.
-    if has_duplicate_keys_in_rmcp_strict_positions(line) {
-        return ValidationOutcome::Reject(invalid_request(Value::Null));
-    }
-    let parsed: Value = match serde_json::from_str(line) {
-        Ok(v) => v,
-        Err(_) => return ValidationOutcome::Reject(parse_error()),
-    };
-    let Some(obj) = parsed.as_object() else {
-        return ValidationOutcome::Reject(invalid_request(Value::Null));
-    };
-    if obj.get("jsonrpc").and_then(Value::as_str) != Some("2.0") {
-        return ValidationOutcome::Reject(invalid_request(extract_id(obj)));
-    }
-
-    // id (if present) must be string|number — null is rejected here
-    // because rmcp's RequestId = NumberOrString won't deserialize null.
-    let id_present_and_valid = match obj.get("id") {
-        None => false,
-        Some(v) if is_forwardable_id(v) => true,
-        Some(_) => return ValidationOutcome::Reject(invalid_request(Value::Null)),
-    };
-
-    let method = obj.get("method");
-    let result = obj.get("result");
-    let error = obj.get("error");
-    let params_ok = obj.get("params").is_none_or(is_valid_params);
-
-    match (method, result, error) {
-        // Request: method+id, no result, no error, valid params shape
-        (Some(m), None, None) if m.is_string() && id_present_and_valid && params_ok => {
-            ValidationOutcome::Forward
-        }
-        // Notification: method, no id, no result, no error, valid params shape
-        (Some(m), None, None) if m.is_string() && !id_present_and_valid && params_ok => {
-            ValidationOutcome::Forward
-        }
-        // Response: id+result, no method, no error
-        (None, Some(_), None) if id_present_and_valid => ValidationOutcome::Forward,
-        // Error response: id+error, no method, no result, error well-formed
-        (None, None, Some(err)) if id_present_and_valid && is_well_formed_error(err) => {
-            ValidationOutcome::Forward
-        }
-        // Catch-all: non-string method, empty object, response/error without
-        // id, malformed error, both result+error, method mixed with
-        // result/error, request/notification with non-structured params.
-        _ => ValidationOutcome::Reject(invalid_request(extract_id(obj))),
-    }
-}
-
-/// Serialize a rejection envelope to a single wire line, terminated
-/// with `\n`. The shape matches JSON-RPC §5: `{jsonrpc, id, error}`
-/// with `error.{code, message}` and no `data`.
-///
-/// **`id` is omitted when null**, not emitted as `"id": null`. MCP's
-/// `RequestId` schema (`integer | string`) disallows null, so emitting
-/// `null` would make the envelope fail the MCP schema validator the
-/// test harness applies via `assert_envelope_valid`. JSON-RPC 2.0 §5
-/// requires `id: null` for parse / invalid-request errors, but MCP's
-/// `JSONRPCErrorResponse` schema marks `id` as OPTIONAL — so omitting
-/// it satisfies both the MCP contract (the protocol we serve) and
-/// the JSON-RPC §5 intent of "no recoverable id available."
-pub(crate) fn synthesize_error_line(env: &ErrorEnvelope) -> String {
-    let body = if env.id.is_null() {
-        serde_json::json!({
-            "jsonrpc": "2.0",
-            "error": {
-                "code": env.code,
-                "message": env.message,
-            },
-        })
-    } else {
-        serde_json::json!({
-            "jsonrpc": "2.0",
-            "id": env.id,
-            "error": {
-                "code": env.code,
-                "message": env.message,
-            },
-        })
-    };
-    // Use to_string (not pretty) so it's exactly one line.
-    let mut line = body.to_string();
-    line.push('\n');
-    line
-}
-
-/// Inbound bridge task. Reads from real stdin one line at a time
-/// (preserving the trailing `\n`), validates each line, and forwards
-/// or rejects.
-///
-/// # Cancellation
-///
-/// This function is intended to be `tokio::spawn`'d and managed by
-/// `ValidatorSupervisor`. If `JoinHandle::abort()` lands between
-/// `write_all` and `flush` on the rejection path, a partial line
-/// may reach real stdout — breaking the stdout serialization
-/// invariant. The supervisor's `shutdown_after_failure` (Task 2.3)
-/// uses abort only on failure paths where stdout integrity is
-/// already lost (e.g. `BrokenPipe`). Cooperative cancellation via
-/// `tokio_util::sync::CancellationToken` could close this gap if
-/// stronger guarantees become necessary; deferred per Task 2.3
-/// design.
-///
-/// Returns `Ok(())` on stdin EOF. Returns `Err(io::Error)` if a write
-/// to either the inbound duplex (forwarding) or shared stdout
-/// (rejection) fails — `BrokenPipe` on stdout is the most common
-/// failure mode and surfaces to `main.rs::run` via the supervisor so
-/// `process_end.reason: Error` is recorded.
-pub(crate) async fn validate_inbound<R>(
-    stdin: R,
-    mut to_rmcp: DuplexStream,
-    stdout: Arc<Mutex<Stdout>>,
-) -> std::io::Result<()>
-where
-    R: AsyncRead + Unpin,
-{
-    let mut reader = BufReader::new(stdin);
-    let mut buf = Vec::with_capacity(BUF_SIZE);
-
-    loop {
-        buf.clear();
-        let n = reader.read_until(b'\n', &mut buf).await?;
-        if n == 0 {
-            return Ok(()); // stdin EOF — clean shutdown.
-        }
-
-        // Strip trailing \n and optional \r for the validation view.
-        // rmcp's codec also strips \r; matching its grammar.
-        // cargo-mutants: known-equivalent group — for the next two
-        // `match` blocks, both `delete match arm Some(&b'\n'/'\r')`
-        // and `replace - with /` mutants are observably equivalent.
-        // The trimmed view is fed to `serde_json::from_str` (and
-        // `Deserializer::from_str` in the dup-check), both of which
-        // tolerate trailing whitespace — so a line that retains its
-        // \n or \r still parses identically. `len() / 1 == len()` so
-        // the slice does not shrink, same effect as deleting the arm.
-        // The `replace - with +` mutant at the \r-strip site IS a
-        // real gap (slices past the buffer end and panics on \r\n
-        // input) and is killed by `validate_inbound_strips_crlf_line_ending`.
-        let trimmed: &[u8] = match buf.last() {
-            Some(&b'\n') => &buf[..buf.len() - 1],
-            _ => &buf[..],
-        };
-        let trimmed: &[u8] = match trimmed.last() {
-            Some(&b'\r') => &trimmed[..trimmed.len() - 1],
-            _ => trimmed,
-        };
-
-        // Non-UTF-8 bytes never reach rmcp. A lossy-decode-then-validate
-        // path would let invalid bytes inside a syntactically valid JSON
-        // string slip through (U+FFFD makes the string parseable), so
-        // the strict check happens before any parsing. JSON-RPC §4.2 / §5.1.
-        let Ok(line_for_validation) = std::str::from_utf8(trimmed) else {
-            let line = synthesize_error_line(&parse_error());
-            let mut sout = stdout.lock().await;
-            sout.write_all(line.as_bytes()).await?;
-            sout.flush().await?;
-            continue;
-        };
-
-        match validate(line_for_validation) {
-            ValidationOutcome::Skip => {}
-            ValidationOutcome::Forward => {
-                // Forward the buffer including its trailing \n so
-                // rmcp's framed reader sees the complete envelope.
-                if buf.last() != Some(&b'\n') {
-                    buf.push(b'\n');
-                }
-                to_rmcp.write_all(&buf).await?;
-                to_rmcp.flush().await?;
-            }
-            ValidationOutcome::Reject(env) => {
-                let line = synthesize_error_line(&env);
-                let mut sout = stdout.lock().await;
-                sout.write_all(line.as_bytes()).await?;
-                sout.flush().await?;
-                // Lock released here; loop continues.
-            }
-        }
-    }
-}
-
-/// Outbound bridge task. Reads newline-framed frames from rmcp's
-/// outbound duplex (rmcp's `FramedWrite` terminates each envelope
-/// with `\n`) and writes each frame through the shared stdout mutex.
-///
-/// Returns `Ok(())` when rmcp drops its outbound duplex end (EOF).
-/// Returns `Err(io::Error)` if writing to real stdout fails —
-/// typically `BrokenPipe`, which surfaces to `main.rs::run` via the
-/// supervisor so `process_end.reason: Error` is recorded.
-///
-/// # Cancellation
-///
-/// Same caveat as `validate_inbound`: an abort between `write_all`
-/// and `flush` may leave a partial frame on stdout. The supervisor's
-/// `shutdown_after_failure` (Task 2.3) uses abort only on failure
-/// paths where stdout integrity is already suspect.
-pub(crate) async fn passthrough_outbound(
-    from_rmcp: DuplexStream,
-    stdout: Arc<Mutex<Stdout>>,
-) -> std::io::Result<()> {
-    let mut reader = BufReader::new(from_rmcp);
-    let mut buf = Vec::with_capacity(BUF_SIZE);
-
-    loop {
-        buf.clear();
-        let n = reader.read_until(b'\n', &mut buf).await?;
-        if n == 0 {
-            return Ok(()); // rmcp dropped outbound — clean shutdown.
-        }
-        let mut sout = stdout.lock().await;
-        sout.write_all(&buf).await?;
-        sout.flush().await?;
-        // Lock released; loop.
-    }
-}
-
-impl ValidatorSupervisor {
-    /// Polls each bridge `JoinHandle` in turn. Resolves with the
-    /// first bridge-task error encountered, OR `Ok(())` once both
-    /// bridges exit `Ok` cleanly (exotic mid-service condition —
-    /// usually one side stays alive until the service ends). Used for
-    /// fail-fast during the `service.waiting()` / `serve_server` race
-    /// in `main.rs::run`.
-    ///
-    /// **Side effect.** When an arm fires, the corresponding
-    /// `JoinHandle` has been polled to completion and its value
-    /// consumed; the `*_consumed` flag is set so the success-path
-    /// `drain` and failure-path `shutdown_after_failure` skip the
-    /// already-polled handle (tokio panics on re-poll of a completed
-    /// `JoinHandle`). The skipped result is treated as `Ok(())` per
-    /// the invariant that this method only returns `Ok` when both
-    /// bridges exited `Ok` (the `Err` arms return early and `drain`
-    /// is not called on the error path).
-    pub async fn watch_for_error(&mut self) -> std::io::Result<()> {
-        loop {
-            tokio::select! {
-                biased;
-                r = &mut self.inbound, if !self.inbound_consumed => {
-                    self.inbound_consumed = true;
-                    match Self::flatten(r) {
-                        Ok(()) if self.outbound_consumed => return Ok(()),
-                        Ok(()) => {}
-                        Err(e) => return Err(e),
-                    }
-                }
-                r = &mut self.outbound, if !self.outbound_consumed => {
-                    self.outbound_consumed = true;
-                    match Self::flatten(r) {
-                        Ok(()) if self.inbound_consumed => return Ok(()),
-                        Ok(()) => {}
-                        Err(e) => return Err(e),
-                    }
-                }
-                else => return Ok(()),
-            }
-        }
-    }
-
-    /// Success-path shutdown. Awaits both bridge tasks; returns the
-    /// first error encountered, else `Ok(())`. Use when
-    /// `service.waiting()` resolved `Ok` (which implies rmcp saw EOF
-    /// on its read, which implies inbound already exited — drain is
-    /// then essentially instant on inbound and bounded on outbound).
-    ///
-    /// **Do not call on failure paths** — the inbound bridge only
-    /// exits on real stdin EOF, but a client may legitimately keep
-    /// stdin open while waiting for the error response, causing this
-    /// to hang. Use `shutdown_after_failure` instead.
-    ///
-    /// **Already-consumed handles are skipped.** If `watch_for_error`
-    /// previously polled a `JoinHandle` to completion, its `_consumed`
-    /// flag is set and the await is skipped (re-polling panics).
-    /// The invariant guarantees the skipped result was `Ok`.
-    pub async fn drain(mut self) -> std::io::Result<()> {
-        let in_r = Self::take_or_await(&mut self.inbound, &mut self.inbound_consumed).await;
-        let out_r = Self::take_or_await(&mut self.outbound, &mut self.outbound_consumed).await;
-        in_r.and(out_r)
-    }
-
-    /// Failure-path shutdown. Aborts the inbound bridge (the client
-    /// may keep stdin open while waiting for an error response;
-    /// without abort, we'd block forever in `read_until` on real
-    /// stdin), then awaits the outbound bridge to drain rmcp's
-    /// queued error envelope plus any validator-synthesized
-    /// rejections. Returns the first error from the outbound path;
-    /// inbound cancellation is expected and ignored.
-    ///
-    /// Used on EVERY failure path — pre-init `ExpectedInitializeRequest`,
-    /// `InitializeFailed`, post-init bridge race error, and post-init
-    /// `service.waiting()` error.
-    ///
-    /// **Already-consumed handles are skipped.** Same contract as
-    /// `drain` — re-polling a completed `JoinHandle` panics, so the
-    /// `_consumed` flags gate the awaits.
-    pub async fn shutdown_after_failure(mut self) -> std::io::Result<()> {
-        // `abort` is always safe (no-op on a finished task); it does
-        // not poll the `JoinHandle` so the consumed flag is irrelevant.
-        self.inbound.abort();
-        // Either it raced to EOF (Ok), got aborted (JoinError), or
-        // was already consumed by `watch_for_error`. We don't care
-        // about the inbound result on this path — the outbound may
-        // still have queued frames to flush.
-        let _ = Self::take_or_await(&mut self.inbound, &mut self.inbound_consumed).await;
-        Self::take_or_await(&mut self.outbound, &mut self.outbound_consumed).await
-    }
-
-    /// Await `handle` if its `consumed` flag is `false`, then mark it
-    /// consumed. Returns `Ok(())` if already consumed — guarded by
-    /// the invariant documented on `drain` and `shutdown_after_failure`.
-    async fn take_or_await(
-        handle: &mut JoinHandle<std::io::Result<()>>,
-        consumed: &mut bool,
-    ) -> std::io::Result<()> {
-        if *consumed {
-            return Ok(());
-        }
-        *consumed = true;
-        Self::flatten(handle.await)
-    }
-
-    fn flatten(r: Result<std::io::Result<()>, tokio::task::JoinError>) -> std::io::Result<()> {
-        match r {
-            Ok(inner) => inner,
-            Err(je) if je.is_cancelled() => Ok(()),
-            Err(je) => Err(std::io::Error::other(format!("bridge task panic: {je}"))),
-        }
-    }
-}
-
-/// Build the validated stdio transport. The two bridge tasks are
-/// spawned immediately; their lifecycle is exposed via `supervisor`.
-/// The returned `transport` plugs into `rmcp::serve_server(server,
-/// validated.transport)` and the returned `stdout` is the shared
-/// writer that `main.rs::run`'s pre-init error emitter must also
-/// lock so all stdout writes (validator rejections, passthrough
-/// frames, pre-init envelopes) serialize through a single mutex.
-///
-/// Drop the returned `transport` ends to signal EOF to rmcp; drop
-/// the supervisor (via `drain`/`shutdown_after_failure`) to await
-/// the bridge tasks' final exits.
-#[must_use]
-pub fn stdio_with_validation() -> ValidatedStdio {
-    let (inbound_our_end, inbound_rmcp_end) = tokio::io::duplex(BUF_SIZE);
-    let (outbound_rmcp_end, outbound_our_end) = tokio::io::duplex(BUF_SIZE);
-
-    let stdout = Arc::new(Mutex::new(tokio::io::stdout()));
-
-    let inbound = tokio::spawn(validate_inbound(
-        tokio::io::stdin(),
-        inbound_our_end,
-        Arc::clone(&stdout),
-    ));
-    let outbound = tokio::spawn(passthrough_outbound(outbound_our_end, Arc::clone(&stdout)));
-
-    ValidatedStdio {
-        transport: (inbound_rmcp_end, outbound_rmcp_end),
-        stdout,
-        supervisor: ValidatorSupervisor {
-            inbound,
-            outbound,
-            inbound_consumed: false,
-            outbound_consumed: false,
-        },
-    }
-}
-
-#[cfg(test)]
-#[expect(
-    clippy::unwrap_used,
-    clippy::expect_used,
-    clippy::panic,
-    reason = "tests"
-)]
-mod tests {
-    use super::*;
-    use serde_json::json;
-
-    fn reject(code: i32, id: Value) -> ValidationOutcome {
-        ValidationOutcome::Reject(ErrorEnvelope {
-            code,
-            message: if code == -32700 {
-                "Parse error"
-            } else {
-                "Invalid Request"
-            },
-            id,
-        })
-    }
-
-    #[test]
-    fn empty_line_skips() {
-        assert_eq!(validate(""), ValidationOutcome::Skip);
-        assert_eq!(validate("   "), ValidationOutcome::Skip);
-        assert_eq!(validate("\r"), ValidationOutcome::Skip);
-    }
-
-    #[test]
-    fn invalid_json_returns_parse_error() {
-        assert_eq!(validate("not valid json"), reject(-32700, Value::Null));
-    }
-
-    #[test]
-    fn non_object_returns_invalid_request() {
-        assert_eq!(validate("[1, 2, 3]"), reject(-32600, Value::Null));
-        assert_eq!(validate("\"a string\""), reject(-32600, Value::Null));
-        assert_eq!(validate("42"), reject(-32600, Value::Null));
-        assert_eq!(validate("null"), reject(-32600, Value::Null));
-    }
-
-    #[test]
-    fn missing_jsonrpc_returns_invalid_request() {
-        assert_eq!(validate(r#"{"method":"a"}"#), reject(-32600, Value::Null));
-    }
-
-    #[test]
-    fn wrong_jsonrpc_value_returns_invalid_request() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"1.0","method":"x","id":1}"#),
-            reject(-32600, json!(1))
-        );
-    }
-
-    #[test]
-    fn duplicate_top_level_keys_reject() {
-        // Cargo-fuzz oracle (#266) caught this libfuzzer-mutated
-        // input. `serde_json::from_str::<Value>` silently collapses
-        // duplicate keys (last wins), so the validator's previous
-        // logic saw a valid Response envelope and forwarded.
-        // rmcp's `serde(untagged)` + `serde(flatten)` machinery
-        // rejects duplicates ("data did not match any variant of
-        // untagged enum JsonRpcMessage"), so the line was silently
-        // dropped — the exact failure mode the validator exists to
-        // prevent. `has_duplicate_keys_in_rmcp_strict_positions`
-        // short-circuits before the dedup happens.
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","id":99,"res":"2.0","id":99,"result":{"x":1}}"#),
-            reject(-32600, Value::Null)
-        );
-    }
-
-    #[test]
-    fn duplicate_keys_inside_error_body_reject() {
-        // Cargo-fuzz oracle (#266) caught a libfuzzer-mutated input
-        // containing duplicate `message` keys inside the `error`
-        // body. rmcp's `ErrorData` is a strict
-        // `#[derive(Deserialize)]` struct that rejects duplicate
-        // fields, so the envelope failed all four variants of
-        // `JsonRpcMessage` and was silently dropped — even though
-        // the top-level keys were unique. The duplicate-key check
-        // recurses one level into `error` to mirror rmcp's strict
-        // deserialize at that position.
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","id":1,"error":{"code":175,"message":75,"message":"x"}}"#),
-            reject(-32600, Value::Null)
-        );
-    }
-
-    #[test]
-    fn duplicate_keys_inside_params_still_forwards() {
-        // Duplicates inside `params` / `result` / `error.data` are
-        // lenient: rmcp deserializes those positions into
-        // `serde_json::Value` (or the equivalent), which preserves
-        // last-wins semantics that match `serde_json::from_str`.
-        // Only top-level keys AND the immediate children of `error`
-        // are checked (rmcp's `ErrorData` struct).
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"x","id":1,"params":{"k":1,"k":2}}"#),
-            ValidationOutcome::Forward
-        );
-    }
-
-    #[test]
-    fn duplicate_keys_inside_error_data_still_forwards() {
-        // Positive control for the recurse-into-error policy:
-        // duplicates inside `error.data` are fine because rmcp's
-        // `ErrorData::data` is `Option<Value>` (lenient).
-        assert_eq!(
-            validate(
-                r#"{"jsonrpc":"2.0","id":1,"error":{"code":-32000,"message":"x","data":{"k":1,"k":2}}}"#
-            ),
-            ValidationOutcome::Forward
-        );
-    }
-
-    #[test]
-    fn typoed_jsonrpc_with_fractional_id_does_not_echo_id() {
-        // Cargo-fuzz oracle (#266) caught this seed input. The
-        // typo'd `.sonrpc` key fails the `jsonrpc == "2.0"` check,
-        // so we route to `extract_id` BEFORE the id-validity branch.
-        // The lax `extract_id` previously echoed `id: 2.5` into the
-        // synthesized envelope, which then failed MCP's `RequestId =
-        // integer | string` schema. `extract_id` is now symmetric
-        // with `is_forwardable_id` so fractional numbers fall
-        // through to Null (and `synthesize_error_line` omits the
-        // id field on Null).
-        assert_eq!(
-            validate(r#"{".sonrpc":"2.0","id":2.5}"#),
-            reject(-32600, Value::Null)
-        );
-    }
-
-    #[test]
-    fn valid_request_forwards() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"tools/list","id":1}"#),
-            ValidationOutcome::Forward
-        );
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"tools/list","id":"abc"}"#),
-            ValidationOutcome::Forward
-        );
-    }
-
-    #[test]
-    fn valid_notification_forwards() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"notifications/cancelled"}"#),
-            ValidationOutcome::Forward
-        );
-    }
-
-    #[test]
-    fn null_id_returns_invalid_request() {
-        // Even though JSON-RPC §5 allows null id on responses, rmcp's
-        // RequestId = NumberOrString rejects it on the wire; we match
-        // that grammar.
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"x","id":null}"#),
-            reject(-32600, Value::Null)
-        );
-    }
-
-    #[test]
-    fn array_or_object_id_returns_invalid_request() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"x","id":[1,2]}"#),
-            reject(-32600, Value::Null)
-        );
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"x","id":{"k":"v"}}"#),
-            reject(-32600, Value::Null)
-        );
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"x","id":true}"#),
-            reject(-32600, Value::Null)
-        );
-    }
-
-    #[test]
-    fn non_string_method_returns_invalid_request() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":42,"id":1}"#),
-            reject(-32600, json!(1))
-        );
-    }
-
-    #[test]
-    fn fractional_id_rejects() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"x","id":1.5}"#),
-            reject(-32600, Value::Null)
-        );
-    }
-
-    #[test]
-    fn out_of_i64_range_id_rejects() {
-        // serde_json may parse very large integers as f64; either way
-        // as_i64() returns None and we reject.
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"x","id":9223372036854775808}"#),
-            reject(-32600, Value::Null)
-        );
-    }
-
-    #[test]
-    fn valid_response_forwards() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","id":99,"result":{"x":1}}"#),
-            ValidationOutcome::Forward
-        );
-    }
-
-    #[test]
-    fn valid_error_response_forwards() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","id":99,"error":{"code":-32601,"message":"not found"}}"#),
-            ValidationOutcome::Forward
-        );
-    }
-
-    #[test]
-    fn response_without_id_returns_invalid_request() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","result":{}}"#),
-            reject(-32600, Value::Null)
-        );
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","error":{"code":1,"message":"x"}}"#),
-            reject(-32600, Value::Null)
-        );
-    }
-
-    #[test]
-    fn both_result_and_error_returns_invalid_request() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","id":1,"result":{},"error":{"code":1,"message":"x"}}"#),
-            reject(-32600, json!(1))
-        );
-    }
-
-    #[test]
-    fn malformed_error_body_returns_invalid_request() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","id":1,"error":{"code":"x","message":"y"}}"#),
-            reject(-32600, json!(1))
-        );
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","id":1,"error":{"message":"only message"}}"#),
-            reject(-32600, json!(1))
-        );
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","id":1,"error":"not even an object"}"#),
-            reject(-32600, json!(1))
-        );
-    }
-
-    #[test]
-    fn fractional_error_code_rejects() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","id":1,"error":{"code":1.5,"message":"x"}}"#),
-            reject(-32600, json!(1))
-        );
-    }
-
-    #[test]
-    fn out_of_i32_range_error_code_rejects() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","id":1,"error":{"code":2147483648,"message":"x"}}"#),
-            reject(-32600, json!(1))
-        );
-    }
-
-    #[test]
-    fn params_as_number_rejects() {
-        // rmcp's codec silently drops non-structured params, making the
-        // server appear to hang; the validator rejects before forwarding.
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"tools/list","id":0,"params":0}"#),
-            reject(-32600, json!(0))
-        );
-    }
-
-    #[test]
-    fn params_as_string_rejects() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"x","id":1,"params":"foo"}"#),
-            reject(-32600, json!(1))
-        );
-    }
-
-    #[test]
-    fn params_as_bool_rejects() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"x","id":1,"params":true}"#),
-            reject(-32600, json!(1))
-        );
-    }
-
-    #[test]
-    fn params_as_array_rejects() {
-        // Cargo-fuzz oracle (#266) caught this seed-corpus input:
-        // `is_valid_params` originally accepted arrays (JSON-RPC §4.2
-        // nominally allows them), but rmcp's `CustomRequest` /
-        // `CustomNotification` deserialize through a `WithMeta`
-        // wrapper using `serde(flatten)`, which fails on non-map
-        // shapes — so an array body causes rmcp to silently drop the
-        // line. Reject in the validator instead. See is_valid_params
-        // docstring for the full chain.
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"x","id":1,"params":[1,2,3]}"#),
-            reject(-32600, json!(1))
-        );
-    }
-
-    #[test]
-    fn params_as_null_forwards() {
-        // rmcp tolerates `params: null` as "no parameters"; matching
-        // that grammar avoids over-rejecting legitimate clients.
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"x","id":1,"params":null}"#),
-            ValidationOutcome::Forward
-        );
-    }
-
-    #[test]
-    fn notification_with_bad_params_rejects() {
-        // §4.1 notifications get the same params shape constraint as
-        // requests.
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"notifications/x","params":0}"#),
-            reject(-32600, Value::Null)
-        );
-    }
-
-    #[test]
-    fn empty_object_returns_invalid_request() {
-        assert_eq!(validate(r"{}"), reject(-32600, Value::Null));
-    }
-
-    #[test]
-    fn mixed_method_and_result_returns_invalid_request() {
-        assert_eq!(
-            validate(r#"{"jsonrpc":"2.0","method":"x","id":1,"result":{}}"#),
-            reject(-32600, json!(1))
-        );
-    }
-
-    #[test]
-    fn synthesize_invalid_request_with_null_id_omits_id_field() {
-        // Pins the MCP-schema accommodation: when the original request
-        // had no recoverable id, the synthesizer OMITS `id` rather than
-        // emitting `"id": null`. MCP's `JSONRPCErrorResponse` schema
-        // marks `id` optional but its type is `integer | string` —
-        // emitting null would fail `assert_envelope_valid` and break
-        // the negative-path wire tests (e.g.
-        // `valid_json_invalid_envelope_returns_minus_32600`).
-        let env = ErrorEnvelope {
-            code: -32600,
-            message: "Invalid Request",
-            id: Value::Null,
-        };
-        let line = synthesize_error_line(&env);
-        assert!(line.ends_with('\n'));
-        let parsed: Value = serde_json::from_str(line.trim_end()).unwrap();
-        assert_eq!(parsed["jsonrpc"], "2.0");
-        let obj = parsed.as_object().unwrap();
-        assert!(
-            !obj.contains_key("id"),
-            "id field MUST be omitted when null; got {parsed}",
-        );
-        assert_eq!(parsed["error"]["code"], -32600);
-        assert_eq!(parsed["error"]["message"], "Invalid Request");
-        // No data field.
-        assert!(parsed["error"].get("data").is_none());
-    }
-
-    #[test]
-    fn synthesize_parse_error_has_correct_code() {
-        let line = synthesize_error_line(&parse_error());
-        let parsed: Value = serde_json::from_str(line.trim_end()).unwrap();
-        assert_eq!(parsed["error"]["code"], -32700);
-        assert_eq!(parsed["error"]["message"], "Parse error");
-    }
-
-    #[test]
-    fn synthesize_echoes_numeric_id() {
-        let env = invalid_request(json!(42));
-        let line = synthesize_error_line(&env);
-        let parsed: Value = serde_json::from_str(line.trim_end()).unwrap();
-        assert_eq!(parsed["id"], 42);
-    }
-
-    #[test]
-    fn synthesize_echoes_string_id() {
-        let env = invalid_request(json!("abc"));
-        let line = synthesize_error_line(&env);
-        let parsed: Value = serde_json::from_str(line.trim_end()).unwrap();
-        assert_eq!(parsed["id"], "abc");
-    }
-
-    use tokio::io::AsyncReadExt;
-
-    #[tokio::test]
-    async fn validate_inbound_forwards_valid_envelope() {
-        let stdin = std::io::Cursor::new(
-            b"{\"jsonrpc\":\"2.0\",\"method\":\"tools/list\",\"id\":1}\n".to_vec(),
-        );
-        let (our_end, mut rmcp_end) = tokio::io::duplex(BUF_SIZE);
-        let stdout = Arc::new(Mutex::new(tokio::io::stdout()));
-
-        // Spawn the validator and consume rmcp's side concurrently.
-        let consumer = tokio::spawn(async move {
-            let mut buf = Vec::new();
-            rmcp_end.read_to_end(&mut buf).await.unwrap();
-            buf
-        });
-        validate_inbound(stdin, our_end, stdout).await.unwrap();
-        let forwarded = consumer.await.unwrap();
-
-        let s = std::str::from_utf8(&forwarded).unwrap();
-        assert!(
-            s.contains("\"method\":\"tools/list\""),
-            "expected forwarded line, got {s:?}",
-        );
-        assert!(s.ends_with('\n'));
-    }
-
-    #[tokio::test]
-    async fn validate_inbound_skips_empty_lines() {
-        let stdin =
-            std::io::Cursor::new(b"\n\n{\"jsonrpc\":\"2.0\",\"method\":\"x\",\"id\":1}\n".to_vec());
-        let (our_end, mut rmcp_end) = tokio::io::duplex(BUF_SIZE);
-        let stdout = Arc::new(Mutex::new(tokio::io::stdout()));
-
-        let consumer = tokio::spawn(async move {
-            let mut buf = Vec::new();
-            rmcp_end.read_to_end(&mut buf).await.unwrap();
-            buf
-        });
-        validate_inbound(stdin, our_end, stdout).await.unwrap();
-        let forwarded = consumer.await.unwrap();
-
-        let s = std::str::from_utf8(&forwarded).unwrap();
-        assert_eq!(s.lines().count(), 1);
-    }
-
-    #[tokio::test]
-    async fn validate_inbound_non_utf8_returns_parse_error() {
-        // Bare non-UTF-8 bytes: the rejection goes to stdout (not
-        // observable from this test) and nothing reaches rmcp.
-        let stdin = std::io::Cursor::new(vec![0xFF, 0xFE, b'\n']);
-        let (our_end, mut rmcp_end) = tokio::io::duplex(BUF_SIZE);
-        let stdout = Arc::new(Mutex::new(tokio::io::stdout()));
-
-        let consumer = tokio::spawn(async move {
-            let mut buf = Vec::new();
-            rmcp_end.read_to_end(&mut buf).await.unwrap();
-            buf
-        });
-        validate_inbound(stdin, our_end, stdout).await.unwrap();
-        let forwarded = consumer.await.unwrap();
-
-        // Nothing reaches rmcp (the invalid-UTF-8 line was rejected
-        // and the rejection went to stdout, not the duplex).
-        assert!(
-            forwarded.is_empty(),
-            "non-UTF-8 line should not be forwarded to rmcp, got {forwarded:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn invalid_utf8_inside_valid_json_string_returns_parse_error() {
-        // Bad UTF-8 inside an otherwise-parseable JSON string literal:
-        // lossy decoding would replace the byte with U+FFFD and pass
-        // validation, leaving the raw byte to reach rmcp. The strict
-        // check rejects before parsing.
-        let mut line: Vec<u8> = br#"{"jsonrpc":"2.0","id":1,"method":"x","params":{"k":""#.to_vec();
-        line.push(0xFF);
-        line.extend_from_slice(br#""}}"#);
-        line.push(b'\n');
-
-        let stdin = std::io::Cursor::new(line);
-        let (our_end, mut rmcp_end) = tokio::io::duplex(BUF_SIZE);
-        let stdout = Arc::new(Mutex::new(tokio::io::stdout()));
-
-        let consumer = tokio::spawn(async move {
-            let mut buf = Vec::new();
-            rmcp_end.read_to_end(&mut buf).await.unwrap();
-            buf
-        });
-        validate_inbound(stdin, our_end, stdout).await.unwrap();
-        let forwarded = consumer.await.unwrap();
-        assert!(
-            forwarded.is_empty(),
-            "bad UTF-8 inside JSON string must NOT be forwarded, got {forwarded:?}",
-        );
-    }
-
-    #[tokio::test]
-    async fn passthrough_outbound_drops_on_eof() {
-        let (rmcp_end, our_end) = tokio::io::duplex(BUF_SIZE);
-        let stdout = Arc::new(Mutex::new(tokio::io::stdout()));
-
-        // Drop the rmcp side immediately — passthrough should see EOF
-        // and return Ok cleanly.
-        drop(rmcp_end);
-        let result = passthrough_outbound(our_end, stdout).await;
-        assert!(result.is_ok(), "expected Ok on EOF, got {result:?}");
-    }
-
-    #[tokio::test]
-    async fn drain_returns_ok_when_both_bridges_exit_ok() {
-        let inbound = tokio::spawn(async { Ok::<(), std::io::Error>(()) });
-        let outbound = tokio::spawn(async { Ok::<(), std::io::Error>(()) });
-        let supervisor = ValidatorSupervisor {
-            inbound,
-            outbound,
-            inbound_consumed: false,
-            outbound_consumed: false,
-        };
-        assert!(supervisor.drain().await.is_ok());
-    }
-
-    #[tokio::test]
-    async fn drain_surfaces_first_error() {
-        let inbound = tokio::spawn(async {
-            Err::<(), std::io::Error>(std::io::Error::other("inbound boom"))
-        });
-        let outbound = tokio::spawn(async { Ok::<(), std::io::Error>(()) });
-        let supervisor = ValidatorSupervisor {
-            inbound,
-            outbound,
-            inbound_consumed: false,
-            outbound_consumed: false,
-        };
-        let r = supervisor.drain().await;
-        assert!(r.is_err());
-        assert!(r.unwrap_err().to_string().contains("inbound boom"));
-    }
-
-    #[tokio::test]
-    async fn shutdown_after_failure_aborts_inbound() {
-        // Inbound never completes on its own — simulates a client
-        // holding stdin open.
-        let inbound = tokio::spawn(async { std::future::pending::<std::io::Result<()>>().await });
-        let outbound = tokio::spawn(async { Ok::<(), std::io::Error>(()) });
-        let supervisor = ValidatorSupervisor {
-            inbound,
-            outbound,
-            inbound_consumed: false,
-            outbound_consumed: false,
-        };
-
-        // Should return promptly (within a few ms) thanks to abort.
-        let r = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            supervisor.shutdown_after_failure(),
-        )
-        .await;
-        assert!(r.is_ok(), "shutdown_after_failure did not abort inbound");
-        assert!(r.unwrap().is_ok());
-    }
-
-    #[tokio::test]
-    async fn watch_for_error_returns_on_first_error() {
-        let inbound = tokio::spawn(async {
-            tokio::time::sleep(std::time::Duration::from_millis(10)).await;
-            Err::<(), std::io::Error>(std::io::Error::other("inbound failed"))
-        });
-        let outbound = tokio::spawn(async { std::future::pending::<std::io::Result<()>>().await });
-        let mut supervisor = ValidatorSupervisor {
-            inbound,
-            outbound,
-            inbound_consumed: false,
-            outbound_consumed: false,
-        };
-
-        let r = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            supervisor.watch_for_error(),
-        )
-        .await;
-        assert!(r.is_ok());
-        let inner = r.unwrap();
-        assert!(inner.is_err());
-        assert!(inner.unwrap_err().to_string().contains("inbound failed"));
-    }
-
-    #[tokio::test]
-    async fn watch_for_error_returns_ok_when_both_bridges_finish() {
-        // Both bridges exit Ok cleanly. watch_for_error must return
-        // Ok(()) without panicking, even if both `is_finished()`
-        // guards evaluate to false in the same select! iteration.
-        let inbound = tokio::spawn(async { Ok::<(), std::io::Error>(()) });
-        let outbound = tokio::spawn(async { Ok::<(), std::io::Error>(()) });
-        // Give the runtime a tick so both handles transition to finished.
-        tokio::task::yield_now().await;
-        let mut supervisor = ValidatorSupervisor {
-            inbound,
-            outbound,
-            inbound_consumed: false,
-            outbound_consumed: false,
-        };
-        let r = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            supervisor.watch_for_error(),
-        )
-        .await
-        .expect("watch_for_error must return promptly when both bridges have finished");
-        assert!(r.is_ok(), "expected Ok, got {r:?}");
-    }
-
-    #[tokio::test]
-    async fn stdio_with_validation_constructs_cleanly() {
-        let validated = stdio_with_validation();
-        // Verify the supervisor is present and we can drop everything.
-        drop(validated.transport);
-        // After dropping the duplex ends, both bridges should exit cleanly
-        // (stdin EOF or read 0 bytes from a dropped duplex).
-        let r = tokio::time::timeout(
-            std::time::Duration::from_secs(2),
-            validated.supervisor.drain(),
-        )
-        .await;
-        // Either Ok (drained cleanly) OR Err (stdin returned an error
-        // on the test runner) is acceptable — we mainly want to confirm
-        // no panic and no hang.
-        assert!(r.is_ok(), "drain hung after dropping transport");
-    }
-
-    #[test]
-    fn error_body_as_primitive_echoes_id() {
-        // OneLevelDupCheck::visit_<primitive> methods must return
-        // Ok(false): a non-map JSON value can have no keys, so no
-        // duplicates are possible. The validate() rejection then comes
-        // from the shape-validation catch-all, which echoes the
-        // original id via extract_id(obj) — NOT Value::Null from the
-        // dup-check rejection path. Each case kills one
-        // OneLevelDupCheck visit-method mutant.
-        let cases: &[(&str, Value, &str)] = &[
-            // Negative integer — serde_json routes through `visit_i64`.
-            // (Positive integers route through `visit_u64` instead, so
-            // this case is what actually exercises the i64 visitor.)
-            (
-                r#"{"jsonrpc":"2.0","id":1,"error":-175}"#,
-                json!(1),
-                "visit_i64",
-            ),
-            // Positive integer fitting in i64 still routes through u64
-            // in serde_json, so this case kills visit_u64.
-            (
-                r#"{"jsonrpc":"2.0","id":2,"error":18446744073709551615}"#,
-                json!(2),
-                "visit_u64",
-            ),
-            (
-                r#"{"jsonrpc":"2.0","id":3,"error":1.5}"#,
-                json!(3),
-                "visit_f64",
-            ),
-            (
-                r#"{"jsonrpc":"2.0","id":4,"error":true}"#,
-                json!(4),
-                "visit_bool",
-            ),
-            (
-                r#"{"jsonrpc":"2.0","id":5,"error":null}"#,
-                json!(5),
-                "visit_unit",
-            ),
-        ];
-        for (line, expected_id, label) in cases {
-            let outcome = validate(line);
-            let ValidationOutcome::Reject(env) = outcome else {
-                panic!("{label}: expected Reject, got {outcome:?}");
-            };
-            assert_eq!(
-                env.id, *expected_id,
-                "{label}: id MUST echo original (not Null); line={line}",
-            );
-            assert_eq!(env.code, -32600, "{label}: expected -32600 invalid-request");
-        }
-    }
-
-    #[tokio::test]
-    async fn validate_inbound_strips_crlf_line_ending() {
-        // Lines terminated with \r\n (typical Windows line endings)
-        // must be handled cleanly: the trailing \n is stripped by the
-        // first match, then the trailing \r is stripped by the second.
-        // The slice arithmetic `&trimmed[..trimmed.len() - 1]` at the
-        // \r-strip site must use subtraction; the `replace - with +`
-        // mutation would slice past the end and panic.
-        let stdin = std::io::Cursor::new(
-            b"{\"jsonrpc\":\"2.0\",\"method\":\"tools/list\",\"id\":1}\r\n".to_vec(),
-        );
-        let (our_end, mut rmcp_end) = tokio::io::duplex(BUF_SIZE);
-        let stdout = Arc::new(Mutex::new(tokio::io::stdout()));
-
-        let consumer = tokio::spawn(async move {
-            let mut buf = Vec::new();
-            rmcp_end.read_to_end(&mut buf).await.unwrap();
-            buf
-        });
-        validate_inbound(stdin, our_end, stdout).await.unwrap();
-        let forwarded = consumer.await.unwrap();
-
-        let s = std::str::from_utf8(&forwarded).unwrap();
-        assert!(
-            s.contains("\"method\":\"tools/list\""),
-            "expected forwarded line, got {s:?}",
-        );
-    }
-
-    #[tokio::test]
-    async fn shutdown_after_failure_surfaces_outbound_error() {
-        // shutdown_after_failure aborts inbound (discarding its
-        // result) and then awaits outbound. The outbound's error MUST
-        // surface as the supervisor's return value — the stub
-        // `Ok(())` mutation would silently swallow this and report
-        // success on every failure path.
-        let inbound = tokio::spawn(async { std::future::pending::<std::io::Result<()>>().await });
-        let outbound = tokio::spawn(async {
-            Err::<(), std::io::Error>(std::io::Error::other("outbound boom"))
-        });
-        let supervisor = ValidatorSupervisor {
-            inbound,
-            outbound,
-            inbound_consumed: false,
-            outbound_consumed: false,
-        };
-        let r = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            supervisor.shutdown_after_failure(),
-        )
-        .await
-        .expect("shutdown_after_failure must not hang");
-        assert!(r.is_err(), "outbound error must surface");
-        assert!(
-            r.unwrap_err().to_string().contains("outbound boom"),
-            "expected outbound error message",
-        );
-    }
-
-    #[tokio::test]
-    async fn watch_for_error_surfaces_bridge_task_panic() {
-        // A panicking bridge task produces a JoinError where
-        // `je.is_cancelled()` is false. The flatten match guard must
-        // NOT collapse all JoinErrors to Ok(()) — only the
-        // cancellation variant. The `replace match guard with true`
-        // mutation would silently swallow panics.
-        let inbound: JoinHandle<std::io::Result<()>> = tokio::spawn(async move {
-            panic!("intentional bridge-task panic for flatten guard test");
-        });
-        let outbound = tokio::spawn(async { std::future::pending::<std::io::Result<()>>().await });
-        let mut supervisor = ValidatorSupervisor {
-            inbound,
-            outbound,
-            inbound_consumed: false,
-            outbound_consumed: false,
-        };
-        let r = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            supervisor.watch_for_error(),
-        )
-        .await
-        .expect("watch_for_error must surface bridge panic promptly");
-        assert!(r.is_err(), "panic must surface as Err, got {r:?}");
-        assert!(
-            r.unwrap_err().to_string().contains("bridge task panic"),
-            "expected bridge-panic error message",
-        );
-    }
-
-    #[tokio::test]
-    async fn watch_for_error_treats_cancellation_as_ok() {
-        // An aborted bridge task produces a cancellation-variant
-        // JoinError. The flatten match guard must treat this as
-        // Ok(()) — abort is the supervisor's own primitive in
-        // shutdown_after_failure and must not be misreported as an
-        // error. The `replace match guard with false` mutation would
-        // surface every cancellation as an Err.
-        let inbound = tokio::spawn(async { std::future::pending::<std::io::Result<()>>().await });
-        inbound.abort();
-        let outbound = tokio::spawn(async { Ok::<(), std::io::Error>(()) });
-        // Give the aborted inbound a tick to transition to finished.
-        tokio::task::yield_now().await;
-        let mut supervisor = ValidatorSupervisor {
-            inbound,
-            outbound,
-            inbound_consumed: false,
-            outbound_consumed: false,
-        };
-        let r = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            supervisor.watch_for_error(),
-        )
-        .await
-        .expect("watch_for_error must return promptly when bridges finish");
-        assert!(
-            r.is_ok(),
-            "cancelled inbound + Ok outbound must yield Ok, got {r:?}",
-        );
-    }
-}
